refactor: simplify pinned memories after code review
- Extract shared countPinnedMemories() to memoryMeta.ts with
  short-circuit at limit, eliminating duplicate counting logic
  in MemoryTool and SettingsViewMessageHandler
- Replace loadMemoryItems() (full preview walk) with lightweight
  countPinnedMemories() in handlePinMemory — avoids double directory
  walk on each pin operation
- Use MemoryPathMessageSchema.extend() for Pin/UnpinMemoryMessageSchema
  instead of duplicating the storagePath field definition
- Remove redundant `?? undefined` on optional-chained expression
- Derive pinned styling from item data via CSS class instead of
  maintaining a separate reflected property that could drift
