+++ src/settingsView/frontend/components/memory/MemoryList.ts
@@ -53,11 +53,7 @@ export class MemoryList extends LitElement {
         ${repeat(
           this.items,
           (item) => item.storagePath,
-          (item) =>
-            html`<memory-item
-              .item=${item}
-              ?pinned=${item.pinned}
-            ></memory-item>`,
+          (item) => html`<memory-item .item=${item}></memory-item>`,
         )}
       </div>
     `;
